+++ server/CHANGES.md
@@ -1,4 +1,11 @@
 # unreleased
+## Remove
+- remove `Builder::listen_unix`
+
+## Change
+- rename `http3` feature to `quic`
+- `Builder::listen` is able to accept generic socket listener
+- update `xitca-io` to `0.3.0`
 
 # 0.2.0
 ## Change