+++ tls/CHANGES.md
@@ -1,4 +1,5 @@
 # unreleased
+- update `xitca-io` to `0.3.0`
 
 # 0.2.3
 ## Add